feat: Add UIComponentFactory Abstract Factory — Standard and Accessible UI families (#4)
