+++ src/creational_patterns/abstract_factory/CourseCard.java
@@ -1,0 +1,5 @@
+package creational_patterns.abstract_factory;
+
+public interface CourseCard {
+    String render(String courseTitle, String category, double price);
+}
