+++ src/creational_patterns/abstract_factory/DashboardCard.java
@@ -1,0 +1,5 @@
+package creational_patterns.abstract_factory;
+
+public interface DashboardCard {
+    String render(String title, String value);
+}
