+++ src/creational_patterns/abstract_factory/UIComponentFactory.java
@@ -1,0 +1,6 @@
+package creational_patterns.abstract_factory;
+
+public interface UIComponentFactory {
+    DashboardCard createDashboardCard();
+    CourseCard    createCourseCard();
+}
